requirements.txt
Flask
konlpy
google-generativeai
pandas
openpyxl
kiwipiepy